requirements.txt
fastapi>=0.109.1
uvicorn>=0.27.0
python-multipart>=0.0.9
pillow>=10.2.0
openai>=1.12.0
google-cloud-vision>=3.8.0
google-cloud-language>=2.14.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0